+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [3.15.0] - 2024-08-05
+### Added
+- option `async`, which allows us to get the result of an expression evaluation
+  asynchronously.
+- Support for asynchronous functions: if any function in an expression returns
+  a Promise and option `async=true`, then the result of evaluating
+  the expression is a Promise.
+- async function `memberOf`.
+
 ## [3.14.1] - 2024-07-02
 ### Fixed
 - impossibility to use attribute name that starts with a capital letter.